+++ CHANGELOG.md
@@ -7,6 +7,14 @@ All Notable changes to `Period` will be documented in this file
 ### Added
 
 - `Duration::withoutCarryOver`
+- `Datepoint::isBefore`
+- `Datepoint::bordersOnStart`
+- `Datepoint::isStarting`
+- `Datepoint::isDuring`
+- `Datepoint::isEnding`
+- `Datepoint::bordersOnEnd`
+- `Datepoint::abuts`
+- `Datepoint::isAfter`
 
 ### Fixed